+++ app/src/main/java/com/exemple/facilita/model/DetalhesValor.kt
@@ -1,9 +1,9 @@
 package com.exemple.facilita.model
 import java.io.Serializable
 data class DetalhesValor(
-    val valor_base: Int,
-    val valor_adicional: Int,
-    val valor_distancia: Int,
-    val valor_total: Int,
+    val valor_base: Double,
+    val valor_adicional: Double,
+    val valor_distancia: Double,
+    val valor_total: Double,
     val detalhes: DetalhesInternos
 ): Serializable
